+++ app/Models/UploadedFile.php
@@ -59,10 +59,6 @@ class UploadedFile extends Model
 
     public function getUrl(): string
     {
-        if ($this->isPublicStored()) {
-            return url(Storage::url($this->path));
-        }
-
         return route('file', $this->id);
     }
 }